feat: move natural language input to SearchFilter component
